% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspr-loading.R
\name{load_screenshot_data}
\alias{load_screenshot_data}
\title{Loads screenshot data from a folder}
\usage{
load_screenshot_data(folder)
}
\arguments{
\item{folder}{}
}
\value{

}
\description{
the function expects a single recording in the folder
}
