% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspr-visualisation.R
\name{plot_positions_and_path}
\alias{plot_positions_and_path}
\title{Plots position and object data}
\usage{
plot_positions_and_path(fspr_obj, i_screenshot = NULL)
}
\arguments{
\item{fspr_obj}{fspr screenshot object}

\item{i_screenshotindex}{index of a screenthot. If provided, it marks
a position on the path where the camera was at the time of that particular screenshot}
}
\value{

}
\description{
Plots position and object data
}
