% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspr-visualisation.R
\name{plot_screenshot_scene}
\alias{plot_screenshot_scene}
\title{plots position of objects on the screen with approximate size as per their
 visual size}
\usage{
plot_screenshot_scene(fspr_obj, i_screenshot, width = 0.1)
}
\arguments{
\item{i_screenshot}{}
}
\value{

}
\description{
plots position of objects on the screen with approximate size as per their
 visual size
}
