% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspr-processing.R
\name{vector3_to_columns}
\alias{vector3_to_columns}
\title{turns vector columns in string "(x, y, z)" into three
columns(position_x, position_y, position_z) and returns the table}
\usage{
vector3_to_columns(df_position, column, remove = TRUE)
}
\description{
turns vector columns in string "(x, y, z)" into three
columns(position_x, position_y, position_z) and returns the table
}
