% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspr-visualisation.R
\name{plot_positions_and_path}
\alias{plot_positions_and_path}
\title{Plots position and object data}
\usage{
plot_positions_and_path(fspr_obj, i_screenshot = NULL)
}
\arguments{
\item{fspr_obj}{}
}
\value{

}
\description{
Plots position and object data
}
