% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspr-visualisation.R
\name{plot_screenshot_scene}
\alias{plot_screenshot_scene}
\title{plots position of objects on the screen with approximate size as per their
 visual size}
\usage{
plot_screenshot_scene(
  fspr_obj,
  i_screenshot,
  include_screenshot = FALSE,
  screenshots_path = NULL
)
}
\arguments{
\item{fspr_obj}{object to plot}

\item{i_screenshot}{which screenshot to plot}

\item{screenshots_path}{}
}
\value{

}
\description{
plots position of objects on the screen with approximate size as per their
 visual size
}
